requirements.txt
pychromecast
plexapi
python-crontab